refactor(orchestrator): extract dashboard settings workflows
