+++ services/local-service/internal/orchestrator/service.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"path"
 	"path/filepath"
-	"reflect"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -73,12 +71,6 @@ type budgetDowngradeDecision struct {
 	DegradeActions []string
 	Summary        string
 	Trace          map[string]any
-}
-
-type modelSecretRollback struct {
-	provider string
-	record   storage.SecretRecord
-	existed  bool
 }
 
 // NewService wires the main orchestration dependencies.
@@ -874,275 +866,6 @@ func (s *Service) NotepadConvertToTask(params map[string]any) (map[string]any, e
 	}, nil
 }
 
-// DashboardOverviewGet handles `agent.dashboard.overview.get`.
-func (s *Service) DashboardOverviewGet(params map[string]any) (map[string]any, error) {
-	queryViews := newTaskQueryViews(s)
-	unfinishedTasks := queryViews.tasks("unfinished", "updated_at", "desc")
-	finishedTasks := queryViews.tasks("finished", "finished_at", "desc")
-	_, runtimePendingTotal := s.runEngine.PendingApprovalRequests(20, 0)
-	needStorageFallback := !queryViews.hasRuntimeState()
-
-	pendingApprovals := pendingApprovalsFromTasks(unfinishedTasks)
-	pendingTotal := mergedPendingApprovalTotal(unfinishedTasks, runtimePendingTotal)
-	focusMode := boolValue(params, "focus_mode", false)
-	requestedIncludes := stringSliceValue(params["include"])
-	includeAll := len(requestedIncludes) == 0
-	includeSet := make(map[string]struct{}, len(requestedIncludes))
-	for _, value := range requestedIncludes {
-		includeSet[value] = struct{}{}
-	}
-
-	focusTask, hasFocusTask := focusTaskForOverview(unfinishedTasks, finishedTasks)
-	var focusSummary map[string]any
-	if hasFocusTask && shouldIncludeOverviewField(includeAll, includeSet, "focus_summary") {
-		focusSummary = map[string]any{
-			"task_id":      focusTask.TaskID,
-			"title":        focusTask.Title,
-			"status":       focusTask.Status,
-			"current_step": focusTask.CurrentStep,
-			"next_action":  nextActionForTask(focusTask),
-			"updated_at":   focusTask.UpdatedAt.Format(dateTimeLayout),
-		}
-	}
-
-	allTasks := append(append([]runengine.TaskRecord{}, unfinishedTasks...), finishedTasks...)
-	hasRestorePoint := latestRestorePointFromTasks(allTasks) != nil
-	if !hasRestorePoint {
-		hasRestorePoint = s.latestRestorePointFromStorage("") != nil
-	}
-	latestAudit := latestAuditRecordFromTasks(allTasks)
-	if latestAudit == nil {
-		latestAudit = s.latestAuditRecordFromStorage("")
-	}
-	quickActions := []string(nil)
-	if shouldIncludeOverviewField(includeAll, includeSet, "quick_actions") {
-		quickActions = buildDashboardQuickActions(hasFocusTask, pendingTotal, len(finishedTasks))
-		if focusMode {
-			quickActions = filterDashboardQuickActionsForFocus(quickActions)
-		}
-	}
-	var globalState map[string]any
-	if shouldIncludeOverviewField(includeAll, includeSet, "global_state") {
-		// Only include global_state when runtime engine has active state
-		// to avoid contradictory data in cold-start fallback scenarios
-		if !needStorageFallback {
-			globalState = s.Snapshot()
-		}
-	}
-	highValueSignal := []string(nil)
-	if shouldIncludeOverviewField(includeAll, includeSet, "high_value_signal") {
-		highValueSignal = buildDashboardSignalsWithAudit(unfinishedTasks, finishedTasks, pendingApprovals, latestAudit)
-		if contextValue := mapValue(params, "context"); len(contextValue) > 0 {
-			highValueSignal = append(highValueSignal, perception.BehaviorSignals(perception.CaptureContextSignals("dashboard", "hover", contextValue))...)
-			highValueSignal = dedupeStringSlice(highValueSignal)
-		}
-		if focusMode {
-			highValueSignal = filterDashboardSignalsForFocus(highValueSignal)
-		}
-	}
-	var trustSummary map[string]any
-	if shouldIncludeOverviewField(includeAll, includeSet, "trust_summary") {
-		trustSummary = map[string]any{
-			"risk_level":             aggregateRiskLevel(allTasks, pendingApprovals, s.risk.DefaultLevel()),
-			"pending_authorizations": pendingTotal,
-			"has_restore_point":      hasRestorePoint,
-			"workspace_path":         currentRuntimeWorkspaceRoot(s.executor),
-		}
-	}
-
-	overview := map[string]any{}
-	if shouldIncludeOverviewField(includeAll, includeSet, "focus_summary") {
-		overview["focus_summary"] = focusSummary
-	} else {
-		overview["focus_summary"] = nil
-	}
-	if shouldIncludeOverviewField(includeAll, includeSet, "trust_summary") {
-		overview["trust_summary"] = trustSummary
-	} else {
-		overview["trust_summary"] = nil
-	}
-	if shouldIncludeOverviewField(includeAll, includeSet, "quick_actions") {
-		overview["quick_actions"] = quickActions
-	} else {
-		overview["quick_actions"] = []string{}
-	}
-	if shouldIncludeOverviewField(includeAll, includeSet, "global_state") {
-		overview["global_state"] = globalState
-	} else {
-		overview["global_state"] = map[string]any{}
-	}
-	if shouldIncludeOverviewField(includeAll, includeSet, "high_value_signal") {
-		overview["high_value_signal"] = highValueSignal
-	} else {
-		overview["high_value_signal"] = []string{}
-	}
-
-	return map[string]any{"overview": overview}, nil
-}
-
-func pendingApprovalsFromTasks(tasks []runengine.TaskRecord) []map[string]any {
-	items := make([]map[string]any, 0, len(tasks))
-	for _, task := range tasks {
-		if task.Status != "waiting_auth" || len(task.ApprovalRequest) == 0 {
-			continue
-		}
-		item := cloneMap(task.ApprovalRequest)
-		if stringValue(item, "task_id", "") == "" {
-			item["task_id"] = task.TaskID
-		}
-		if stringValue(item, "risk_level", "") == "" {
-			item["risk_level"] = task.RiskLevel
-		}
-		items = append(items, item)
-	}
-	return items
-}
-
-func approvalRequestRecordsToItems(records []storage.ApprovalRequestRecord) []map[string]any {
-	items := make([]map[string]any, 0, len(records))
-	for _, record := range records {
-		item := map[string]any{
-			"approval_id":    record.ApprovalID,
-			"task_id":        record.TaskID,
-			"operation_name": record.OperationName,
-			"risk_level":     record.RiskLevel,
-			"target_object":  record.TargetObject,
-			"reason":         record.Reason,
-			"status":         record.Status,
-			"created_at":     record.CreatedAt,
-			"updated_at":     record.UpdatedAt,
-		}
-		if strings.TrimSpace(record.ImpactScopeJSON) != "" {
-			var scope map[string]any
-			if err := json.Unmarshal([]byte(record.ImpactScopeJSON), &scope); err == nil && len(scope) > 0 {
-				item["impact_scope"] = scope
-			}
-		}
-		items = append(items, item)
-	}
-	return items
-}
-
-// mergedPendingApprovalTotal prefers the task-centric merged view so mixed
-// runtime and storage snapshots report one stable pending-authorization count.
-func mergedPendingApprovalTotal(unfinishedTasks []runengine.TaskRecord, runtimePendingTotal int) int {
-	pendingTotal := countPendingApprovalTasks(unfinishedTasks)
-	if pendingTotal == 0 && runtimePendingTotal > 0 {
-		return runtimePendingTotal
-	}
-	return pendingTotal
-}
-
-// DashboardModuleGet handles `agent.dashboard.module.get`.
-func (s *Service) DashboardModuleGet(params map[string]any) (map[string]any, error) {
-	module := stringValue(params, "module", "mirror")
-	tab := stringValue(params, "tab", "daily_summary")
-	queryViews := newTaskQueryViews(s)
-	finishedTasks := queryViews.tasks("finished", "finished_at", "desc")
-	unfinishedTasks := queryViews.tasks("unfinished", "updated_at", "desc")
-	_, runtimePendingTotal := s.runEngine.PendingApprovalRequests(20, 0)
-	pendingTotal := mergedPendingApprovalTotal(unfinishedTasks, runtimePendingTotal)
-	latestAudit := latestAuditRecordFromTasks(append(append([]runengine.TaskRecord{}, unfinishedTasks...), finishedTasks...))
-	if latestAudit == nil {
-		latestAudit = s.latestAuditRecordFromStorage("")
-	}
-	pluginSummary := s.pluginRuntimeSummary()
-	summary := map[string]any{
-		"completed_tasks":     len(finishedTasks),
-		"generated_outputs":   countGeneratedOutputs(finishedTasks),
-		"authorizations_used": countAuthorizedTasks(unfinishedTasks, finishedTasks),
-		"exceptions":          countExceptionTasks(unfinishedTasks, finishedTasks),
-		"plugin_runtime":      pluginSummary,
-	}
-	highlights := buildDashboardModuleHighlightsWithAudit(unfinishedTasks, finishedTasks, pendingTotal, latestAudit)
-	if module == "tasks" {
-		summary = s.buildDashboardTaskModuleSummary(unfinishedTasks, finishedTasks, summary)
-		highlights = s.buildDashboardTaskModuleHighlights(unfinishedTasks, finishedTasks, pendingTotal, latestAudit)
-	}
-	return map[string]any{
-		"module":     module,
-		"tab":        tab,
-		"summary":    summary,
-		"highlights": highlights,
-	}, nil
-}
-
-// buildDashboardTaskModuleSummary keeps the generic dashboard module summary
-// while exposing one task-focused runtime summary for the current focus task.
-func (s *Service) buildDashboardTaskModuleSummary(unfinishedTasks, finishedTasks []runengine.TaskRecord, baseSummary map[string]any) map[string]any {
-	summary := cloneMap(baseSummary)
-	summary["processing_tasks"] = countTasksWithStatus(unfinishedTasks, "processing")
-	summary["waiting_auth_tasks"] = countTasksWithStatus(unfinishedTasks, "waiting_auth")
-	summary["blocked_tasks"] = countTasksWithStatus(unfinishedTasks, "blocked", "failed", "ended_unfinished", "paused")
-	focusTask, ok := focusTaskForOverview(unfinishedTasks, finishedTasks)
-	if !ok {
-		return summary
-	}
-	summary["focus_task_id"] = focusTask.TaskID
-	summary["focus_runtime_summary"] = s.buildDashboardFocusRuntimeSummary(focusTask)
-	return summary
-}
-
-// buildDashboardTaskModuleHighlights turns the current focus task runtime into
-// human-readable dashboard hints without adding a new protocol method.
-func (s *Service) buildDashboardTaskModuleHighlights(unfinishedTasks, finishedTasks []runengine.TaskRecord, pendingTotal int, latestAudit map[string]any) []string {
-	highlights := make([]string, 0, 6)
-	focusTask, ok := focusTaskForOverview(unfinishedTasks, finishedTasks)
-	if ok {
-		runtimeSummary := s.buildDashboardFocusRuntimeSummary(focusTask)
-		if focusTask.Status == "waiting_auth" {
-			highlights = append(highlights, "焦点任务当前正在等待授权确认。")
-		} else if focusTask.Status == "processing" {
-			highlights = append(highlights, fmt.Sprintf("焦点任务仍在执行中，当前步骤为 %s。", firstNonEmptyString(focusTask.CurrentStep, "generate_output")))
-		} else if focusTask.Status == "blocked" || focusTask.Status == "failed" || focusTask.Status == "paused" || focusTask.Status == "ended_unfinished" {
-			highlights = append(highlights, fmt.Sprintf("焦点任务当前状态为 %s。", focusTask.Status))
-		}
-		if stopReason := strings.TrimSpace(stringValue(runtimeSummary, "loop_stop_reason", "")); stopReason != "" {
-			highlights = append(highlights, fmt.Sprintf("最近停止原因：%s。", stopReason))
-		}
-		if latestEventType := strings.TrimSpace(stringValue(runtimeSummary, "latest_event_type", "")); latestEventType != "" {
-			highlights = append(highlights, fmt.Sprintf("最近运行事件：%s。", latestEventType))
-		}
-		if steeringCount := intValue(runtimeSummary, "active_steering_count", 0); steeringCount > 0 {
-			highlights = append(highlights, fmt.Sprintf("当前仍有 %d 条追加要求待消费。", steeringCount))
-		}
-	}
-	highlights = append(highlights, buildDashboardModuleHighlightsWithAudit(unfinishedTasks, finishedTasks, pendingTotal, latestAudit)...)
-	return dedupeStringSlice(highlights)
-}
-
-// buildDashboardFocusRuntimeSummary reuses the task detail runtime summary but
-// allows dashboard cards to fall back to the latest in-memory runtime event
-// when persistence has not yet flushed a loop event row.
-func (s *Service) buildDashboardFocusRuntimeSummary(task runengine.TaskRecord) map[string]any {
-	summary := s.buildTaskRuntimeSummary(task)
-	if strings.TrimSpace(stringValue(summary, "latest_event_type", "")) != "" {
-		return summary
-	}
-	latestEventType := strings.TrimSpace(stringValue(task.LatestEvent, "type", ""))
-	if strings.HasPrefix(latestEventType, "loop.") || latestEventType == "task.steered" {
-		summary["latest_event_type"] = latestEventType
-	}
-	return summary
-}
-
-// MirrorOverviewGet handles `agent.mirror.overview.get`.
-func (s *Service) MirrorOverviewGet(params map[string]any) (map[string]any, error) {
-	_ = params
-	finishedTasks := newTaskQueryViews(s).tasks("finished", "finished_at", "desc")
-	memoryReferences := collectMirrorReferences(finishedTasks)
-	return map[string]any{
-		"history_summary": buildMirrorHistorySummary(finishedTasks, memoryReferences),
-		"daily_summary": map[string]any{
-			"date":              time.Now().Format("2006-01-02"),
-			"completed_tasks":   len(finishedTasks),
-			"generated_outputs": countGeneratedOutputs(finishedTasks),
-		},
-		"profile":           buildMirrorProfile(finishedTasks),
-		"memory_references": memoryReferences,
-	}, nil
-}
-
 // PluginRuntimeList exposes the smallest backend query surface for runtime
 // plugin visibility so dashboard work can consume health and metric snapshots
 // without depending on static worker declarations only.
@@ -1669,143 +1392,6 @@ func (s *Service) SecurityRespond(params map[string]any) (map[string]any, error)
 	return response, nil
 }
 
-// SettingsGet handles agent.settings.get.
-func (s *Service) SettingsGet(params map[string]any) (map[string]any, error) {
-	settings := normalizeSettingsSnapshot(s.runEngine.Settings())
-	scope := normalizeSettingsScope(stringValue(params, "scope", "all"))
-	if scope == "all" || scope == "models" {
-		settingsWithSecrets, err := s.attachSensitiveSettingAvailability(settings)
-		if err != nil {
-			return nil, err
-		}
-		settings = settingsWithSecrets
-	}
-	if scope == "all" {
-		return map[string]any{"settings": settings}, nil
-	}
-
-	section, ok := settings[scope].(map[string]any)
-	if !ok {
-		return map[string]any{"settings": map[string]any{}}, nil
-	}
-
-	return map[string]any{"settings": map[string]any{scope: cloneMap(section)}}, nil
-}
-
-// SettingsUpdate handles agent.settings.update and returns the effective
-// settings patch plus apply-mode metadata.
-func (s *Service) SettingsUpdate(params map[string]any) (map[string]any, error) {
-	normalizedParams := normalizeSettingsUpdateParams(params)
-	previewSettings, previewUpdatedKeys, _, _, err := s.previewSettingsUpdate(normalizedParams)
-	if err != nil {
-		return nil, err
-	}
-	modelSettingsChanged := modelSettingsTouched(previewUpdatedKeys)
-	modelSecretTouched := false
-	secretUpdatedKeys := make([]string, 0, 2)
-	rollbacks := make([]modelSecretRollback, 0, 2)
-	previousModel := s.currentModel()
-	if models := cloneMap(mapValue(normalizedParams, "models")); len(models) > 0 {
-		if deleteAPIKey := boolValue(models, "delete_api_key", false); deleteAPIKey {
-			provider := s.providerForSettingsUpdate(models)
-			rollback, rollbackErr := s.captureModelSecretRollback(provider)
-			if rollbackErr != nil {
-				return nil, rollbackErr
-			}
-			if err := s.deleteModelSecret(provider); err != nil {
-				return nil, err
-			}
-			rollbacks = append(rollbacks, rollback)
-			delete(models, "delete_api_key")
-			normalizedParams["models"] = models
-			modelSecretTouched = true
-			secretUpdatedKeys = append(secretUpdatedKeys, "models.delete_api_key")
-		}
-		if apiKey := stringValue(models, "api_key", ""); apiKey != "" {
-			provider := s.providerForSettingsUpdate(models)
-			rollback, rollbackErr := s.captureModelSecretRollback(provider)
-			if rollbackErr != nil {
-				return nil, rollbackErr
-			}
-			if err := s.persistModelSecret(provider, apiKey); err != nil {
-				return nil, err
-			}
-			rollbacks = append(rollbacks, rollback)
-			delete(models, "api_key")
-			normalizedParams["models"] = models
-			modelSecretTouched = true
-			secretUpdatedKeys = append(secretUpdatedKeys, "models.api_key")
-		}
-	}
-	if modelSettingsChanged {
-		if err := s.reloadRuntimeModelForSettings(previewSettings); err != nil {
-			s.rollbackModelSecretMutations(rollbacks)
-			return nil, err
-		}
-	}
-	effectiveSettings, updatedKeys, applyMode, needRestart, err := s.runEngine.UpdateSettings(normalizedParams)
-	if err != nil {
-		s.ReplaceModel(previousModel)
-		s.rollbackModelSecretMutations(rollbacks)
-		return nil, err
-	}
-	if modelSettingsChanged {
-		applyMode = "next_task_effective"
-		needRestart = false
-	}
-	if modelSecretTouched {
-		if _, ok := effectiveSettings["models"]; !ok {
-			effectiveSettings["models"] = map[string]any{}
-		}
-	}
-	if _, ok := effectiveSettings["models"]; ok {
-		effectiveSettings = s.attachSensitiveSettingAvailabilityForCommittedUpdate(effectiveSettings, secretUpdatedKeys)
-	}
-	effectiveSettings = outwardSettingsUpdatePatch(effectiveSettings)
-	updatedKeys = outwardSettingsUpdateKeys(updatedKeys, secretUpdatedKeys)
-	return map[string]any{
-		"updated_keys":       updatedKeys,
-		"effective_settings": effectiveSettings,
-		"apply_mode":         applyMode,
-		"need_restart":       needRestart,
-	}, nil
-}
-
-// attachSensitiveSettingAvailabilityForCommittedUpdate decorates the response
-// payload after a settings update has already committed. At this point the
-// runtime model, secrets, and settings snapshot may already be live, so a
-// follow-up Stronghold read must not turn the committed save back into an RPC
-// error. When the readonly secret probe fails, the response degrades to stable
-// metadata derived from the just-applied mutation hints and current Stronghold
-// descriptor instead of reopening a partial-apply path.
-func (s *Service) attachSensitiveSettingAvailabilityForCommittedUpdate(settings map[string]any, secretUpdatedKeys []string) map[string]any {
-	decorated, err := s.attachSensitiveSettingAvailability(settings)
-	if err == nil {
-		return decorated
-	}
-	return attachSensitiveSettingAvailabilityFallback(settings, strongholdStatusFromStorage(s.storage), settingsUpdateSecretAvailabilityHint(secretUpdatedKeys))
-}
-
-// previewSettingsUpdate computes the future settings snapshot without mutating
-// runengine state so SettingsUpdate can validate runtime model reloads before
-// persisting a next-task-effective model route.
-func (s *Service) previewSettingsUpdate(values map[string]any) (map[string]any, []string, string, bool, error) {
-	if s == nil || s.runEngine == nil {
-		return nil, nil, "", false, nil
-	}
-	currentSettings := s.runEngine.Settings()
-	nextSettings := cloneMap(currentSettings)
-	if nextSettings == nil {
-		nextSettings = map[string]any{}
-	}
-	previewPatch := cloneMap(values)
-	mergeSettingsPreview(nextSettings, previewPatch)
-	updatedKeys := settingsPatchPathsFromPreview(previewPatch)
-	applyMode := previewApplyMode(currentSettings, previewPatch, updatedKeys)
-	needRestart := previewNeedsRestart(currentSettings, previewPatch)
-	return normalizeSettingsSnapshot(nextSettings), updatedKeys, applyMode, needRestart, nil
-}
-
 func (s *Service) executeScreenAnalysisAfterApproval(task runengine.TaskRecord, pendingExecution map[string]any) (runengine.TaskRecord, map[string]any, map[string]any, error) {
 	if s.executor == nil || s.executor.ScreenClient() == nil {
 		failedTask, failureBubble := s.failExecutionTask(task, map[string]any{"name": "screen_analyze"}, execution.Result{}, tools.ErrScreenCaptureNotSupported)
@@ -1940,262 +1526,6 @@ func cleanupExpiredScreenTemps(screenClient tools.ScreenCaptureClient, reason st
 	})
 }
 
-func (s *Service) attachSensitiveSettingAvailability(settings map[string]any) (map[string]any, error) {
-	cloned := normalizeSettingsSnapshot(cloneMap(settings))
-	if cloned == nil {
-		cloned = map[string]any{}
-	}
-	models := cloneMap(mapValue(cloned, "models"))
-	if models == nil {
-		models = map[string]any{}
-	}
-	credentials := cloneMap(mapValue(models, "credentials"))
-	if credentials == nil {
-		credentials = map[string]any{}
-	}
-	provider, configured, err := s.modelSecretConfigured(providerFromSettings(models, s.defaultSettingsProvider()))
-	if err != nil {
-		return nil, err
-	}
-	if stringValue(models, "provider", "") == "" && provider != "" {
-		models["provider"] = provider
-	}
-	credentials["provider_api_key_configured"] = configured
-	if stronghold := strongholdStatusFromStorage(s.storage); len(stronghold) > 0 {
-		credentials["stronghold"] = stronghold
-	}
-	models["credentials"] = credentials
-	cloned["models"] = models
-	return cloned, nil
-}
-
-func (s *Service) modelSecretConfigured(provider string) (string, bool, error) {
-	resolvedProvider := model.CanonicalProviderName(firstNonEmptyString(strings.TrimSpace(provider), s.defaultSettingsProvider()))
-	if s.storage == nil || s.storage.SecretStore() == nil || resolvedProvider == "" {
-		return resolvedProvider, false, nil
-	}
-	_, err := s.storage.SecretStore().GetSecret(context.Background(), "model", resolvedProvider+"_api_key")
-	if err == nil {
-		return resolvedProvider, true, nil
-	}
-	if errors.Is(err, storage.ErrSecretNotFound) {
-		return resolvedProvider, false, nil
-	}
-	if errors.Is(err, storage.ErrSecretStoreAccessFailed) {
-		return resolvedProvider, false, ErrStrongholdAccessFailed
-	}
-	if errors.Is(err, storage.ErrStrongholdUnavailable) {
-		return resolvedProvider, false, ErrStrongholdAccessFailed
-	}
-	return resolvedProvider, false, err
-}
-
-func attachSensitiveSettingAvailabilityFallback(settings map[string]any, stronghold map[string]any, providerConfigured *bool) map[string]any {
-	cloned := normalizeSettingsSnapshot(cloneMap(settings))
-	if cloned == nil {
-		cloned = map[string]any{}
-	}
-	models := cloneMap(mapValue(cloned, "models"))
-	if models == nil {
-		models = map[string]any{}
-	}
-	credentials := cloneMap(mapValue(models, "credentials"))
-	if credentials == nil {
-		credentials = map[string]any{}
-	}
-	if providerConfigured != nil {
-		credentials["provider_api_key_configured"] = *providerConfigured
-	} else if _, ok := credentials["provider_api_key_configured"]; !ok {
-		credentials["provider_api_key_configured"] = false
-	}
-	if len(stronghold) > 0 {
-		credentials["stronghold"] = cloneMap(stronghold)
-	}
-	models["credentials"] = credentials
-	cloned["models"] = models
-	return cloned
-}
-
-func settingsUpdateSecretAvailabilityHint(secretUpdatedKeys []string) *bool {
-	for _, key := range secretUpdatedKeys {
-		switch key {
-		case "models.api_key":
-			configured := true
-			return &configured
-		case "models.delete_api_key":
-			configured := false
-			return &configured
-		}
-	}
-	return nil
-}
-
-func (s *Service) persistModelSecret(provider, apiKey string) error {
-	resolvedProvider := model.CanonicalProviderName(firstNonEmptyString(strings.TrimSpace(provider), s.defaultSettingsProvider()))
-	if s.storage == nil || s.storage.SecretStore() == nil || resolvedProvider == "" {
-		return ErrStrongholdAccessFailed
-	}
-	if err := s.storage.SecretStore().PutSecret(context.Background(), storage.SecretRecord{
-		Namespace: "model",
-		Key:       resolvedProvider + "_api_key",
-		Value:     strings.TrimSpace(apiKey),
-		UpdatedAt: time.Now().UTC().Format(time.RFC3339),
-	}); err != nil {
-		normalizedErr := storage.NormalizeSecretStoreError(err)
-		if errors.Is(normalizedErr, storage.ErrStrongholdAccessFailed) || errors.Is(normalizedErr, storage.ErrStrongholdUnavailable) || errors.Is(normalizedErr, storage.ErrSecretStoreAccessFailed) {
-			return ErrStrongholdAccessFailed
-		}
-		return normalizedErr
-	}
-	return nil
-}
-
-func (s *Service) deleteModelSecret(provider string) error {
-	resolvedProvider := model.CanonicalProviderName(firstNonEmptyString(strings.TrimSpace(provider), s.defaultSettingsProvider()))
-	if s.storage == nil || s.storage.SecretStore() == nil || resolvedProvider == "" {
-		return ErrStrongholdAccessFailed
-	}
-	if err := s.storage.SecretStore().DeleteSecret(context.Background(), "model", resolvedProvider+"_api_key"); err != nil {
-		normalizedErr := storage.NormalizeSecretStoreError(err)
-		if errors.Is(normalizedErr, storage.ErrStrongholdAccessFailed) || errors.Is(normalizedErr, storage.ErrStrongholdUnavailable) || errors.Is(normalizedErr, storage.ErrSecretStoreAccessFailed) {
-			return ErrStrongholdAccessFailed
-		}
-		return normalizedErr
-	}
-	return nil
-}
-
-func (s *Service) captureModelSecretRollback(provider string) (modelSecretRollback, error) {
-	resolvedProvider := model.CanonicalProviderName(firstNonEmptyString(strings.TrimSpace(provider), s.defaultSettingsProvider()))
-	rollback := modelSecretRollback{provider: resolvedProvider}
-	if s.storage == nil || s.storage.SecretStore() == nil || resolvedProvider == "" {
-		return rollback, nil
-	}
-	record, err := s.storage.SecretStore().GetSecret(context.Background(), "model", resolvedProvider+"_api_key")
-	if err == nil {
-		rollback.record = record
-		rollback.existed = true
-		return rollback, nil
-	}
-	normalizedErr := storage.NormalizeSecretStoreError(err)
-	if errors.Is(normalizedErr, storage.ErrSecretNotFound) {
-		return rollback, nil
-	}
-	if errors.Is(normalizedErr, storage.ErrStrongholdAccessFailed) {
-		return rollback, ErrStrongholdAccessFailed
-	}
-	return rollback, normalizedErr
-}
-
-func (s *Service) rollbackModelSecretMutations(rollbacks []modelSecretRollback) {
-	for index := len(rollbacks) - 1; index >= 0; index-- {
-		rollback := rollbacks[index]
-		if rollback.provider == "" || s == nil || s.storage == nil || s.storage.SecretStore() == nil {
-			continue
-		}
-		if rollback.existed {
-			_ = s.storage.SecretStore().PutSecret(context.Background(), rollback.record)
-			continue
-		}
-		_ = s.storage.SecretStore().DeleteSecret(context.Background(), "model", rollback.provider+"_api_key")
-	}
-}
-
-func (s *Service) reloadRuntimeModelForSettings(settings map[string]any) error {
-	if s == nil || s.runEngine == nil {
-		return nil
-	}
-	resolvedConfig := model.RuntimeConfigFromSettings(s.currentModelConfig(), settings)
-	modelService, err := model.NewServiceFromConfig(model.ServiceConfig{
-		ModelConfig:  resolvedConfig,
-		SecretSource: model.NewStaticSecretSource(s.storage),
-	})
-	if err != nil {
-		if shouldFallbackRuntimeModelReload(err) {
-			modelService = model.NewService(resolvedConfig)
-		} else {
-			return err
-		}
-	}
-	s.ReplaceModel(modelService)
-	return nil
-}
-
-func settingsPatchPathsFromPreview(patch map[string]any) []string {
-	if len(patch) == 0 {
-		return nil
-	}
-	keys := make([]string, 0, len(patch))
-	for key := range patch {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-	paths := make([]string, 0, len(keys))
-	for _, key := range keys {
-		nextPrefix := key
-		if nested, ok := patch[key].(map[string]any); ok && len(nested) > 0 {
-			for _, child := range settingsPatchPathsFromPreview(nested) {
-				paths = append(paths, nextPrefix+"."+child)
-			}
-			continue
-		}
-		paths = append(paths, nextPrefix)
-	}
-	return paths
-}
-
-func mergeSettingsPreview(target map[string]any, patch map[string]any) {
-	for key, value := range patch {
-		patchMap, ok := value.(map[string]any)
-		if ok {
-			currentMap, currentOK := target[key].(map[string]any)
-			if !currentOK {
-				currentMap = map[string]any{}
-			}
-			mergeSettingsPreview(currentMap, patchMap)
-			target[key] = currentMap
-			continue
-		}
-		target[key] = value
-	}
-}
-
-func previewNeedsRestart(currentSettings, patch map[string]any) bool {
-	generalPatch := cloneMap(mapValue(patch, "general"))
-	if len(generalPatch) > 0 {
-		nextLanguage, ok := generalPatch["language"]
-		if ok {
-			currentGeneral := cloneMap(mapValue(currentSettings, "general"))
-			currentLanguage, hasCurrentLanguage := currentGeneral["language"]
-			if !hasCurrentLanguage || !reflect.DeepEqual(currentLanguage, nextLanguage) {
-				return true
-			}
-		}
-		downloadPatch := cloneMap(mapValue(generalPatch, "download"))
-		if len(downloadPatch) > 0 {
-			nextWorkspacePath, ok := downloadPatch["workspace_path"]
-			if ok {
-				currentDownload := cloneMap(mapValue(mapValue(currentSettings, "general"), "download"))
-				currentWorkspacePath, hasCurrentWorkspacePath := currentDownload["workspace_path"]
-				if !hasCurrentWorkspacePath || !reflect.DeepEqual(currentWorkspacePath, nextWorkspacePath) {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
-func previewApplyMode(currentSettings, patch map[string]any, updatedKeys []string) string {
-	if previewNeedsRestart(currentSettings, patch) {
-		return "restart_required"
-	}
-	if modelSettingsTouched(updatedKeys) {
-		return "next_task_effective"
-	}
-	return "immediate"
-}
-
 func inspectorConfigFromSettings(settings map[string]any) map[string]any {
 	taskAutomation := cloneMap(mapValue(normalizeSettingsSnapshot(settings), "task_automation"))
 	if taskAutomation == nil {
@@ -2323,195 +1653,6 @@ func optionalStringSliceValue(rawValue any) ([]string, bool) {
 	}
 }
 
-func strongholdStatusFromStorage(store *storage.Service) map[string]any {
-	if store == nil || store.Stronghold() == nil {
-		return map[string]any{
-			"backend":      "none",
-			"available":    false,
-			"fallback":     false,
-			"initialized":  false,
-			"formal_store": false,
-		}
-	}
-	descriptor := store.Stronghold().Descriptor()
-	return map[string]any{
-		"backend":      descriptor.Backend,
-		"available":    descriptor.Available,
-		"fallback":     descriptor.Fallback,
-		"initialized":  descriptor.Initialized,
-		"formal_store": descriptor.Available && !descriptor.Fallback,
-	}
-}
-
-func normalizeSettingsScope(scope string) string {
-	switch strings.TrimSpace(scope) {
-	case "", "all":
-		return "all"
-	case "data_log":
-		return "models"
-	default:
-		return strings.TrimSpace(scope)
-	}
-}
-
-func normalizeSettingsSnapshot(settings map[string]any) map[string]any {
-	cloned := cloneMap(settings)
-	if cloned == nil {
-		return map[string]any{}
-	}
-	models := cloneMap(mapValue(cloned, "models"))
-	if models == nil {
-		models = map[string]any{}
-	}
-	if legacy := cloneMap(mapValue(cloned, "data_log")); len(legacy) > 0 {
-		for key, value := range legacy {
-			if key == "provider" {
-				models[key] = value
-				continue
-			}
-			credentials := cloneMap(mapValue(models, "credentials"))
-			if credentials == nil {
-				credentials = map[string]any{}
-			}
-			credentials[key] = value
-			models["credentials"] = credentials
-		}
-		delete(cloned, "data_log")
-	}
-	models = normalizeModelSettingsSection(models)
-	if len(models) > 0 {
-		cloned["models"] = models
-	}
-	return cloned
-}
-
-func normalizeSettingsUpdateParams(params map[string]any) map[string]any {
-	cloned := cloneMap(params)
-	if cloned == nil {
-		return map[string]any{}
-	}
-	models := cloneMap(mapValue(cloned, "models"))
-	if models == nil {
-		models = map[string]any{}
-	}
-	if legacy := cloneMap(mapValue(cloned, "data_log")); len(legacy) > 0 {
-		for key, value := range legacy {
-			if key == "provider_api_key_configured" || key == "stronghold" {
-				continue
-			}
-			models[key] = value
-		}
-		delete(cloned, "data_log")
-	}
-	if credentials := cloneMap(mapValue(models, "credentials")); len(credentials) > 0 {
-		for key, value := range credentials {
-			if key == "provider_api_key_configured" || key == "stronghold" {
-				continue
-			}
-			models[key] = value
-		}
-		delete(models, "credentials")
-	}
-	if len(models) > 0 {
-		cloned["models"] = models
-	}
-	return cloned
-}
-
-func normalizeModelSettingsSection(models map[string]any) map[string]any {
-	cloned := cloneMap(models)
-	if cloned == nil {
-		cloned = map[string]any{}
-	}
-	credentials := cloneMap(mapValue(cloned, "credentials"))
-	if credentials == nil {
-		credentials = map[string]any{}
-	}
-	for _, key := range []string{"budget_auto_downgrade", "base_url", "model", "budget_policy"} {
-		if value, ok := cloned[key]; ok {
-			credentials[key] = value
-			delete(cloned, key)
-		}
-	}
-	if len(credentials) > 0 {
-		cloned["credentials"] = credentials
-	}
-	return cloned
-}
-
-func modelSettingsSection(settings map[string]any) map[string]any {
-	return cloneMap(mapValue(normalizeSettingsSnapshot(settings), "models"))
-}
-
-func modelCredentialSettings(settings map[string]any) map[string]any {
-	return cloneMap(mapValue(modelSettingsSection(settings), "credentials"))
-}
-
-func outwardSettingsUpdatePatch(settings map[string]any) map[string]any {
-	cloned := normalizeSettingsSnapshot(settings)
-	models := cloneMap(mapValue(cloned, "models"))
-	if len(models) == 0 {
-		return cloned
-	}
-	credentials := cloneMap(mapValue(models, "credentials"))
-	delete(models, "credentials")
-	for _, key := range []string{"budget_auto_downgrade", "provider_api_key_configured", "base_url", "model", "stronghold"} {
-		if value, ok := credentials[key]; ok {
-			models[key] = value
-		}
-	}
-	cloned["models"] = models
-	return cloned
-}
-
-func outwardSettingsUpdateKeys(internalKeys, secretUpdatedKeys []string) []string {
-	seen := make(map[string]struct{}, len(internalKeys)+len(secretUpdatedKeys))
-	result := make([]string, 0, len(internalKeys)+len(secretUpdatedKeys))
-	for _, key := range internalKeys {
-		mapped := key
-		if strings.HasPrefix(mapped, "models.credentials.") {
-			mapped = "models." + strings.TrimPrefix(mapped, "models.credentials.")
-		}
-		if _, ok := seen[mapped]; ok {
-			continue
-		}
-		seen[mapped] = struct{}{}
-		result = append(result, mapped)
-	}
-	for _, key := range secretUpdatedKeys {
-		if _, ok := seen[key]; ok {
-			continue
-		}
-		seen[key] = struct{}{}
-		result = append(result, key)
-	}
-	sort.Strings(result)
-	return result
-}
-
-func (s *Service) providerForSettingsUpdate(models map[string]any) string {
-	merged := modelSettingsSection(s.runEngine.Settings())
-	if merged == nil {
-		merged = map[string]any{}
-	}
-	for key, value := range normalizeModelSettingsSection(models) {
-		merged[key] = value
-	}
-	return providerFromSettings(merged, s.defaultSettingsProvider())
-}
-
-func (s *Service) defaultSettingsProvider() string {
-	if s.currentModel() == nil {
-		return ""
-	}
-	return strings.TrimSpace(s.currentModel().Provider())
-}
-
-func providerFromSettings(models map[string]any, fallback string) string {
-	provider := firstNonEmptyString(stringValue(models, "provider", ""), fallback)
-	return model.CanonicalProviderName(provider)
-}
-
 // defaultIntentMap creates a minimal default intent payload for notepad
 // conversions.
 func defaultIntentMap(name string) map[string]any {
@@ -2542,283 +1683,6 @@ func notepadIntent(item map[string]any) map[string]any {
 		return defaultIntentMap("explain")
 	default:
 		return defaultIntentMap("summarize")
-	}
-}
-
-func focusTaskForOverview(unfinishedTasks, finishedTasks []runengine.TaskRecord) (runengine.TaskRecord, bool) {
-	if len(unfinishedTasks) > 0 {
-		return unfinishedTasks[0], true
-	}
-	if len(finishedTasks) > 0 {
-		return finishedTasks[0], true
-	}
-	return runengine.TaskRecord{}, false
-}
-
-func nextActionForTask(task runengine.TaskRecord) string {
-	switch task.Status {
-	case "confirming_intent":
-		return "确认当前意图"
-	case "waiting_auth":
-		return "处理待授权操作"
-	case "waiting_input":
-		return "补充输入内容"
-	case "processing":
-		return "等待处理完成"
-	case "completed":
-		return "查看交付结果"
-	default:
-		return "打开任务详情"
-	}
-}
-
-func buildDashboardQuickActions(hasFocusTask bool, pendingTotal, finishedCount int) []string {
-	actions := make([]string, 0, 3)
-	if pendingTotal > 0 {
-		actions = append(actions, "处理待授权操作")
-	}
-	if hasFocusTask {
-		actions = append(actions, "打开任务详情")
-	}
-	if finishedCount > 0 {
-		actions = append(actions, "查看最近结果")
-	}
-	if len(actions) == 0 {
-		actions = append(actions, "等待新任务")
-	}
-	return actions
-}
-
-func shouldIncludeOverviewField(includeAll bool, includeSet map[string]struct{}, field string) bool {
-	if includeAll {
-		return true
-	}
-	_, ok := includeSet[field]
-	return ok
-}
-
-func filterDashboardQuickActionsForFocus(actions []string) []string {
-	filtered := make([]string, 0, len(actions))
-	for _, action := range actions {
-		if action == "查看最近结果" {
-			continue
-		}
-		filtered = append(filtered, action)
-	}
-	if len(filtered) == 0 {
-		return []string{"打开任务详情"}
-	}
-	return filtered
-}
-
-func filterDashboardSignalsForFocus(signals []string) []string {
-	if len(signals) <= 2 {
-		return signals
-	}
-	return append([]string(nil), signals[:2]...)
-}
-
-func dedupeStringSlice(values []string) []string {
-	seen := make(map[string]struct{}, len(values))
-	result := make([]string, 0, len(values))
-	for _, value := range values {
-		trimmed := strings.TrimSpace(value)
-		if trimmed == "" {
-			continue
-		}
-		if _, ok := seen[trimmed]; ok {
-			continue
-		}
-		seen[trimmed] = struct{}{}
-		result = append(result, trimmed)
-	}
-	return result
-}
-
-func buildDashboardSignals(unfinishedTasks, finishedTasks []runengine.TaskRecord, pendingApprovals []map[string]any) []string {
-	signals := make([]string, 0, 3)
-	if len(unfinishedTasks) > 0 {
-		signals = append(signals, fmt.Sprintf("当前有 %d 个未完成任务处于 runtime 管控中。", len(unfinishedTasks)))
-	}
-	if len(pendingApprovals) > 0 {
-		signals = append(signals, fmt.Sprintf("当前有 %d 个待授权操作等待用户确认。", len(pendingApprovals)))
-	}
-	if latestRestorePointFromTasks(finishedTasks) != nil {
-		signals = append(signals, "最近一次正式交付已经生成可回放的恢复点。")
-	}
-	if len(signals) == 0 {
-		signals = append(signals, "主链路当前暂无活跃任务。")
-	}
-	return signals
-}
-
-func buildDashboardModuleHighlights(unfinishedTasks, finishedTasks []runengine.TaskRecord, pendingTotal int) []string {
-	highlights := make([]string, 0, 4)
-	if latestOutputPath := latestOutputPathFromTasks(finishedTasks); latestOutputPath != "" {
-		highlights = append(highlights, fmt.Sprintf("最近正式交付已落到 %s。", latestOutputPath))
-	}
-	if pendingTotal > 0 {
-		highlights = append(highlights, fmt.Sprintf("当前仍有 %d 个待授权任务等待处理。", pendingTotal))
-	}
-	if restorePoint := latestRestorePointFromTasks(finishedTasks); restorePoint != nil {
-		highlights = append(highlights, fmt.Sprintf("最近恢复点 %s 已可用于安全回显。", stringValue(restorePoint, "recovery_point_id", "latest")))
-	}
-	if len(unfinishedTasks) > 0 {
-		highlights = append(highlights, fmt.Sprintf("最近活跃任务状态为 %s。", unfinishedTasks[0].Status))
-	}
-	if len(highlights) == 0 {
-		highlights = append(highlights, "当前模块视图已切换为 runtime 聚合结果。")
-	}
-	return highlights
-}
-
-func countGeneratedOutputs(tasks []runengine.TaskRecord) int {
-	total := 0
-	for _, task := range tasks {
-		if len(task.DeliveryResult) > 0 || len(task.Artifacts) > 0 {
-			total++
-		}
-	}
-	return total
-}
-
-func buildDashboardSignalsWithAudit(unfinishedTasks, finishedTasks []runengine.TaskRecord, pendingApprovals []map[string]any, latestAudit map[string]any) []string {
-	signals := buildDashboardSignals(unfinishedTasks, finishedTasks, pendingApprovals)
-	if latestAudit != nil {
-		signals = append(signals, fmt.Sprintf("最近审计摘要：%s。", truncateText(stringValue(latestAudit, "summary", "runtime audit recorded"), 48)))
-	}
-	return signals
-}
-
-func buildDashboardModuleHighlightsWithAudit(unfinishedTasks, finishedTasks []runengine.TaskRecord, pendingTotal int, latestAudit map[string]any) []string {
-	highlights := buildDashboardModuleHighlights(unfinishedTasks, finishedTasks, pendingTotal)
-	if latestAudit != nil {
-		highlights = append(highlights, fmt.Sprintf("最近审计动作：%s -> %s。", truncateText(stringValue(latestAudit, "action", "audit"), 24), truncateText(stringValue(latestAudit, "target", "main_flow"), 36)))
-	}
-	return highlights
-}
-
-func countAuthorizedTasks(taskGroups ...[]runengine.TaskRecord) int {
-	total := 0
-	for _, tasks := range taskGroups {
-		for _, task := range tasks {
-			if len(task.Authorization) > 0 {
-				total++
-			}
-		}
-	}
-	return total
-}
-
-func countTasksWithStatus(tasks []runengine.TaskRecord, statuses ...string) int {
-	if len(statuses) == 0 {
-		return 0
-	}
-	allowed := make(map[string]struct{}, len(statuses))
-	for _, status := range statuses {
-		if strings.TrimSpace(status) == "" {
-			continue
-		}
-		allowed[status] = struct{}{}
-	}
-	total := 0
-	for _, task := range tasks {
-		if _, ok := allowed[task.Status]; ok {
-			total++
-		}
-	}
-	return total
-}
-
-func countExceptionTasks(taskGroups ...[]runengine.TaskRecord) int {
-	total := 0
-	for _, tasks := range taskGroups {
-		for _, task := range tasks {
-			switch task.Status {
-			case "failed", "cancelled", "blocked", "ended_unfinished":
-				total++
-			}
-		}
-	}
-	return total
-}
-
-func collectMirrorReferences(tasks []runengine.TaskRecord) []map[string]any {
-	references := make([]map[string]any, 0)
-	seen := map[string]struct{}{}
-	for _, task := range tasks {
-		for _, reference := range task.MirrorReferences {
-			memoryID := stringValue(reference, "memory_id", "")
-			if memoryID == "" {
-				continue
-			}
-			if _, ok := seen[memoryID]; ok {
-				continue
-			}
-			seen[memoryID] = struct{}{}
-			references = append(references, cloneMap(reference))
-		}
-	}
-	return references
-}
-
-func buildMirrorHistorySummary(tasks []runengine.TaskRecord, memoryReferences []map[string]any) []string {
-	if len(tasks) == 0 {
-		return []string{"当前还没有完成任务，镜像概览会在首个正式交付后生成。"}
-	}
-
-	summaries := []string{
-		fmt.Sprintf("最近已完成 %d 个任务，其中 %d 个产出了正式交付。", len(tasks), countGeneratedOutputs(tasks)),
-	}
-	if len(memoryReferences) > 0 {
-		summaries = append(summaries, fmt.Sprintf("当前累计挂接了 %d 条记忆引用，可供 task detail 与 mirror 回显复用。", len(memoryReferences)))
-	}
-	if latestOutputPath := latestOutputPathFromTasks(tasks); latestOutputPath != "" {
-		summaries = append(summaries, fmt.Sprintf("最近一次落盘结果位于 %s。", latestOutputPath))
-	}
-	return summaries
-}
-
-func buildMirrorProfile(tasks []runengine.TaskRecord) map[string]any {
-	if len(tasks) == 0 {
-		return nil
-	}
-
-	documentCount := 0
-	bubbleCount := 0
-	earliestHour := 24
-	latestHour := -1
-	for _, task := range tasks {
-		switch stringValue(task.DeliveryResult, "type", "") {
-		case "workspace_document":
-			documentCount++
-		case "bubble":
-			bubbleCount++
-		}
-		hour := task.StartedAt.Hour()
-		if hour < earliestHour {
-			earliestHour = hour
-		}
-		if hour > latestHour {
-			latestHour = hour
-		}
-	}
-
-	workStyle := "偏好即时结果回显"
-	preferredOutput := "bubble"
-	if documentCount >= bubbleCount {
-		workStyle = "偏好结构化落盘输出"
-		preferredOutput = "workspace_document"
-	}
-	if earliestHour == 24 || latestHour == -1 {
-		earliestHour = 0
-		latestHour = 0
-	}
-
-	return map[string]any{
-		"work_style":       workStyle,
-		"preferred_output": preferredOutput,
-		"active_hours":     fmt.Sprintf("%02d-%02dh", earliestHour, latestHour+1),
 	}
 }
 
@@ -2920,50 +1784,6 @@ func (s *Service) latestToolCallFromStorage(taskID, runID string) map[string]any
 		"error_code":   item.ErrorCode,
 		"duration_ms":  item.DurationMS,
 	}
-}
-
-func aggregateTokenCostSummary(unfinishedTasks, finishedTasks []runengine.TaskRecord, budgetAutoDowngrade bool) map[string]any {
-	currentTaskTokens := 0
-	currentTaskCost := 0.0
-	if currentTask, ok := latestTokenUsageTask(unfinishedTasks, finishedTasks); ok {
-		currentTaskTokens = intValueFromAny(currentTask.TokenUsage["total_tokens"])
-		currentTaskCost = floatValueFromAny(currentTask.TokenUsage["estimated_cost"])
-	}
-
-	todayTokens := 0
-	todayCost := 0.0
-	now := time.Now()
-	for _, task := range append(append([]runengine.TaskRecord{}, unfinishedTasks...), finishedTasks...) {
-		if !sameDay(task.StartedAt, now) {
-			continue
-		}
-		todayTokens += intValueFromAny(task.TokenUsage["total_tokens"])
-		todayCost += floatValueFromAny(task.TokenUsage["estimated_cost"])
-	}
-
-	return map[string]any{
-		"current_task_tokens":   currentTaskTokens,
-		"current_task_cost":     currentTaskCost,
-		"today_tokens":          todayTokens,
-		"today_cost":            todayCost,
-		"single_task_limit":     0.0,
-		"daily_limit":           0.0,
-		"budget_auto_downgrade": budgetAutoDowngrade,
-	}
-}
-
-func latestTokenUsageTask(unfinishedTasks, finishedTasks []runengine.TaskRecord) (runengine.TaskRecord, bool) {
-	for _, task := range unfinishedTasks {
-		if len(task.TokenUsage) > 0 {
-			return task, true
-		}
-	}
-	for _, task := range finishedTasks {
-		if len(task.TokenUsage) > 0 {
-			return task, true
-		}
-	}
-	return runengine.TaskRecord{}, false
 }
 
 func parseAuditTime(auditRecord map[string]any) time.Time {
